# Speed up DangerousGetRowSpan(y)