--- a/src/ImageSharp/Common/Helpers/Numerics.cs
+++ b/src/ImageSharp/Common/Helpers/Numerics.cs
@@ -978,9 +978,9 @@ namespace SixLabors.ImageSharp
         /// Tells whether input value is outside of the 0..max range.
         /// </summary>
         /// <param name="value">Value.</param>
-        /// <param name="max">Maximum value, inclusive.</param>
+        /// <param name="max">Maximum value, exclusive.</param>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool IsOutOfRangeZeroToMax(int value, int max)
-            => (uint)value > (uint)max;
+        public static bool IsOutOfRangeZeroToExclusiveMax(int value, int max)
+            => (uint)value >= (uint)max;
     }
 }