--- a/src/ImageSharp/Common/Helpers/Numerics.cs
+++ b/src/ImageSharp/Common/Helpers/Numerics.cs
@@ -973,14 +973,5 @@ namespace SixLabors.ImageSharp
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool IsOutOfRange(int value, int min, int max)
             => (uint)(value - min) > (uint)(max - min);
-
-        /// <summary>
-        /// Tells whether input value is outside of the 0..max range.
-        /// </summary>
-        /// <param name="value">Value.</param>
-        /// <param name="max">Maximum value, exclusive.</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool IsOutOfRangeZeroToExclusiveMax(int value, int max)
-            => (uint)value >= (uint)max;
     }
 }