… tests/ImageSharp.Benchmarks/General/Buffer2D_DangerousGetRowSpan.cs
… tests/ImageSharp.Tests/Memory/Buffer2DTests.SwapOrCopyContent.cs
… tests/ImageSharp.Tests/Memory/Buffer2DTests.cs
… tests/ImageSharp.Tests/Memory/DiscontiguousBuffers/MemoryGroupTests.SwapOrCopyContent.cs
… tests/ImageSharp.Tests/Memory/DiscontiguousBuffers/MemoryGroupTests.cs