### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Fixes #1884 by caching span data for known `IMemoryOwner` implementations backing `MemoryGroup`.

Additionally, it moves `SwapOrCopyContents` responsibility entirely to `Buffer2D`. Originally I did this because I wanted multiple `MemoryGroup.Owned` subclasses to deal with different `GetRowSpanCoreUnsafe` implementations, but in the end I decided to switch on an enum for simplicity & better perf. Regardless, it's cleaner to swap the memory groups, backing `Buffer2D` instead of swapping their contents.

### Benchmarks - DangerousGetRowSpan

#### Before

```
|              Method | SizeMegaBytes |     Mean |    Error |   StdDev |
|-------------------- |-------------- |---------:|---------:|---------:|
| DangerousGetRowSpan |           0.5 | 75.03 ns | 1.511 ns | 2.352 ns |
| DangerousGetRowSpan |             2 | 72.42 ns | 1.456 ns | 1.618 ns |
| DangerousGetRowSpan |            10 | 73.64 ns | 1.370 ns | 2.640 ns |
```

#### After

```
|              Method | SizeMegaBytes |      Mean |     Error |    StdDev |
|-------------------- |-------------- |----------:|----------:|----------:|
| DangerousGetRowSpan |           0.5 |  8.017 ns | 0.1790 ns | 0.2131 ns |
| DangerousGetRowSpan |             2 |  6.445 ns | 0.1479 ns | 0.1644 ns |
| DangerousGetRowSpan |            10 | 46.089 ns | 0.9179 ns | 0.9015 ns |
```

### Benchmarks - DecodeJpeg

#### Before

```
|                              Method |      Mean |     Error |    StdDev |
|------------------------------------ |----------:|----------:|----------:|
|        'Baseline 4:4:4 Interleaved' | 15.814 ms | 0.2987 ms | 0.2794 ms |
|        'Baseline 4:2:0 Interleaved' | 10.336 ms | 0.2018 ms | 0.2243 ms |
|        'Baseline 4:0:0 (grayscale)' |  2.044 ms | 0.0246 ms | 0.0230 ms |
| 'Progressive 4:2:0 Non-Interleaved' | 13.868 ms | 0.2624 ms | 0.2455 ms |
```

#### After

```
|                              Method |      Mean |     Error |    StdDev |
|------------------------------------ |----------:|----------:|----------:|
|        'Baseline 4:4:4 Interleaved' | 12.682 ms | 0.2425 ms | 0.2695 ms |
|        'Baseline 4:2:0 Interleaved' |  9.241 ms | 0.1263 ms | 0.0986 ms |
|        'Baseline 4:0:0 (grayscale)' |  1.677 ms | 0.0136 ms | 0.0107 ms |
| 'Progressive 4:2:0 Non-Interleaved' | 13.639 ms | 0.2601 ms | 0.3289 ms |
```

/cc @br3aker